+++ sm/lib/options/options_interface.c
@@ -11,7 +11,7 @@
 #include "options.h"
 
 /** User-friendly interface */
-struct option* options_allocate(size_t n) {
+struct option* options_allocate(int n) {
 	struct option* ops = malloc(sizeof(struct option)*n);
 	size_t i; for(i=0;i<n;i++) {
 		ops[i].name = 0;
